Moved point to plane out of the bloat as well
